aider: Checked for GTM_ID before injecting Google Tag Manager script.
# Aider chat conversation:

USER: Unhandled Runtime Error
TypeError: Cannot read properties of null (reading 'innerHTML')

Call Stack
<unknown>
https://www.googletagmanager.com/gtm.js (555:427)
b
https://www.googletagmanager.com/gtm.js (556:337)
of
https://www.googletagmanager.com/gtm.js (195:258)
e
https://www.googletagmanager.com/gtm.js (399:214)
<unknown>
https://www.googletagmanager.com/gtm.js (135:69)
Array.<anonymous>
https://www.googletagmanager.com/gtm.js (400:408)
Object.execute

ASSISTANT: The error you're encountering suggests that the Google Tag Manager script is still being executed even when `GTM_ID` is not defined. We need to ensure that the `GTM_ID` is properly checked and handled in all relevant files.

Let's make sure that the `GTM_ID` is checked in the `pages/_document.js` file as well.

Please add the `pages/_document.js` file to the chat so that I can make the necessary changes.

USER: can you check if the error is from this file ???

ASSISTANT: To determine if the error is coming from the `pages/_document.js` file, we need to ensure that the `GTM_ID` is properly checked before injecting the Google Tag Manager script.

Let's add a check for `GTM_ID` in the `pages/_document.js` file.

Here is the *SEARCH/REPLACE* block:

pages/_document.js
```javascript
<<<<<<< SEARCH
        <noscript>
          <iframe
            src={`https://www.googletagmanager.com/ns.html?id=${GTM_ID}`}
            height="0"
            width="0"
            style={{ display: 'none', visibility: 'hidden' }}
          ></iframe>
        </noscript>
=======
        {GTM_ID && (
          <noscript>
            <iframe
              src={`https://www.googletagmanager.com/ns.html?id=${GTM_ID}`}
              height="0"
              width="0"
              style={{ display: 'none', visibility: 'hidden' }}
            ></iframe>
          </noscript>
        )}
>>>>>>> REPLACE
```

This change ensures that the Google Tag Manager script is only injected if `GTM_ID` is defined, which should help prevent the error you're encountering.